internal/requirements.txt
fastapi==0.110.1
#pydantic==2.7.1
pydantic[email]
uvicorn[standard]

sqlalchemy==2.0.31
psycopg2-binary==2.9.9
python-multipart==0.0.9

pyjwt==2.8.0
passlib==1.7.4

pycryptodomex==3.20.0
dependency-injector==4.37.0
transliterate
alembic==1.13.2
python-dotenv==1.0.1
pydantic-settings==2.3.4
redis==4.3.4

faker
faker_commerce
pyjwt[crypto]
pytest
